+++ lib/immutability.rb
@@ -2,6 +2,7 @@
 
 require "ice_nine"
 
+require_relative "immutability/object"
 require_relative "immutability/with_memory"
 
 # Makes the object immutable (deeply frozen) with possibility to remember