+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.0.2 to be released
+
+### Added
+
+- New decorator `Immutabitlity::Object` to iterate via past states of instance (nepalez)
+- New method `Imutability::WithMemory#at` to return a state in the past (nepalez)
+
+[Compare v0.0.1...HEAD](https://github.com/nepalez/immutability/compare/v0.0.1...HEAD)
+
 ## v0.0.1 2015-09-20
 
 First public release.